# module/player_module/character_handler/character_request/create_character_request.go
package character_request

type CreateCharacter struct {
	Name    string `json:"name" validate:"required"`
	RaceId  string `json:"raceId" validate:"required"`
	ClassId string `json:"classId" validate:"required"`
}